# perf(encode): use Number.isInteger