`Number.isInteger` is a faster operation than flooring a value and then checking it with `isFinite`.

| | tiny │ small │ medium │ large |
|--|--|--|--|--|
| before │ 2,659,084 ops/sec │ 579,367 ops/sec │ 46,217 ops/sec │ 488 ops/sec   |
| after │ 2,784,261 ops/sec │ 618,960 ops/sec │ 46,302 ops/sec │ 480 ops/sec   |