@zdm 2021-01-15
Benchmarks says that Number.isInteger() is slower.